% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.cv.R
\name{bw.cv}
\alias{bw.cv}
\title{Perform bandwith selection by cross-validation for the kernel regression estimator.}
\usage{
bw.cv(data)
}
\arguments{
\item{data}{List of curves to estimate by kernel regression}
}
\value{
A vector containing the mimimum, the median, the mean and the maximum
bandwith selected by cross-validation for every curve.
}
\description{
Perform bandwith selection by cross-validation for the kernel regression estimator.
}
