% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cai2list}
\alias{cai2list}
\title{Convert \code{ssfcov} objects into right format for this package}
\usage{
cai2list(time, x, subject)
}
\arguments{
\item{time}{Vector, observation times.}

\item{x}{Vector, observation values.}

\item{subject}{Vector, observation id.}
}
\value{
A list, where each element represents a curve. Each curve is defined
as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}
}
\description{
Convert \code{ssfcov} objects into right format for this package
}
\references{
ssfcov, R package. See Cai and Yuan, Nonparametric Covariance
Function Estimation for Functional and Longitudinal Data - 2010.
}
