% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_moment.R
\name{covariance}
\alias{covariance}
\title{Perform an estimation of the covariance}
\usage{
covariance(
  data,
  U = seq(0, 1, length.out = 101),
  t0_list = 0.5,
  k0_list = 2,
  t0_list_mean = seq(0.1, 0.9, by = 0.1),
  centered = FALSE,
  dense = FALSE
)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{A vector of numerics, sampling points at which estimate the curves.}

\item{t0_list}{A vector of numerics, the sampling points at which we estimate
\eqn{H0}. We will consider the \eqn{8k0 - 7} nearest points of \eqn{t_0} for
the estimation of \eqn{H_0} when \eqn{\sigma} is unknown.}

\item{k0_list}{A vector of numerics, the number of neighbors of \eqn{t_0} to
consider. Should be set as \deqn{k0 = M* exp(-log(log(M))^2)}. We can set a
different \eqn{k_0}, but in order to use the same for each \eqn{t_0}, just
put a unique numeric.}

\item{centered}{Boolean, default=FALSE. Does the data have already been
centered?}
}
\value{
A matrix of shape (length(U), length(U)) which is an estimation of
the covariance matrix.
}
\description{
This function performs the estimation of the covariance of a set of curves.
}
