% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_b.R
\name{estimate.b.cv}
\alias{estimate.b.cv}
\title{Perform the estimation of the bandwith using CV.}
\usage{
estimate.b.cv(data)
}
\arguments{
\item{data}{List of curves to estimate by kernel regression.}
}
\value{
An estimation of the bandwidth by cross-validation.
}
\description{
Perform the estimation of the bandwith using CV.
}
