% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.b.R
\name{estimate.b.list}
\alias{estimate.b.list}
\title{Perform the estimation of the bandwidth over a list of H0 and t0.}
\usage{
estimate.b.list(data, H0_list, L0_list, sigma = 0, K = "epanechnikov")
}
\arguments{
\item{data}{List of curves to estimate by kernel regression.}

\item{H0_list}{List of estimates of H0.}

\item{L0_list}{<- List of estimates of L0.}

\item{sigma}{An estimation of sigma.}

\item{K}{Which kernel to use?
\itemize{
\item epanechnikov
\item beta
\item uniform
}

@return List of estimates of the bandwidth.
@export}
}
\description{
Perform the estimation of the bandwidth over a list of H0 and t0.
}
