% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.b.R
\name{estimate.b}
\alias{estimate.b}
\title{Perform the estimation of the bandwith}
\usage{
estimate.b(data, H0 = 0.5, L0 = 1, sigma = 0, K = "epanechnikov")
}
\arguments{
\item{data}{List of curves to estimate by kernel regression.}

\item{H0}{An estimation of H0.}

\item{L0}{An estimation of L0.}

\item{sigma}{An estimation of sigma.}

\item{kernel}{Which kernel to use?
\itemize{
\item epanechnikov (default)
\item beta
\item uniform
}}
}
\value{
An estimation of H0.
}
\description{
Perform the estimation of the bandwith
}
