% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.bw.R
\name{estimate.bw}
\alias{estimate.bw}
\title{Perform bandwith selection by estimation of regularity of the curves.}
\usage{
estimate.bw(data, pct = 0.1, sigma = 0, l = 2)
}
\arguments{
\item{data}{List of curves to estimate by kernel regression.}

\item{pct}{Percentage of data to consider for the estimation.}

\item{sigma}{The value of the standard deviation of the noise (estimated or not).}

\item{l}{The sampling time to consider (> 1).}
}
\value{
An estimation of the bandwith
}
\description{
Perform bandwith selection by estimation of regularity of the curves.
}
