% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_curve.R
\name{estimate.curve}
\alias{estimate.curve}
\title{Perform the smoothing of individuals curves.}
\usage{
estimate.curve(curve, U, b, t0_list = NULL, kernel = "epanechnikov")
}
\arguments{
\item{curve}{List of two vectors representing a curve:
- $t Sampling points
- $x Observed points}

\item{U}{Vector of points for the estimation}

\item{b}{Bandwith}

\item{t0_list}{List of time at which the bandwidths have been estimated.}

\item{kernel}{Kernel name to use
\itemize{
\item epanechnikov (default)
\item uniforma
\item beta
\item mBeta
\item locPoly (not used)
}}
}
\value{
List of two vectors representating the estimated curve:
- $t Sampling points
- $x Predicted points
}
\description{
Perform the smoothing of individuals curves.
}
