% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.k0.R
\name{estimate.k0.pilot}
\alias{estimate.k0.pilot}
\title{Perform the estimation of the pilot k0.}
\usage{
estimate.k0.pilot(M)
}
\arguments{
\item{M}{Mean number of samping points per curve.}
}
\value{
The pilot k0.
}
\description{
Perform the estimation of the pilot k0.
}
