% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.L0.R
\name{estimate.L0.list}
\alias{estimate.L0.list}
\title{Perform the estimation of L0 over a list of t0.}
\usage{
estimate.L0.list(data, t0_list, H0_list, k0 = 2, sigma = NULL, density = NULL)
}
\arguments{
\item{data}{List of curves to estimate by kernel regression.}

\item{t0_list}{Starting times for the estimation of H0. We consider the 8k0 - 7
nearest points of t0 for the estimation of H0 when sigma is unknown.}

\item{H0_list}{Estimation of H0 at each t0.}

\item{k0}{For the computation of the gap between the different observations.}

\item{sigma}{True value of sigma.
If null, change estimate.}

\item{density}{Density of the sampling points (currently, only consider
uniform sampling points)?}
}
\value{
A list containing the estimation of H0 at each t0.
}
\description{
Perform the estimation of L0 over a list of t0.
}
