% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.L0.R
\name{estimate.L0}
\alias{estimate.L0}
\title{Perform the estimation of L0.}
\usage{
estimate.L0(data, t0 = 0, H0 = 0, k0 = 2, sigma = NULL, density = NULL)
}
\arguments{
\item{data}{List of curves to estimate by kernel regression.}

\item{t0}{The starting time for the estimation of H0. We consider the 8k0 - 7
nearest points of t0 for the estimation of H0 when sigma is unknown.}

\item{H0}{An estimation of H0.}

\item{k0}{For the computation of the gap between the different observations.
Should be set as k0 = M / max(8, log(M)).}

\item{sigma}{True value of sigma.
If null, change estimate.}

\item{density}{Density of the sampling points (currently, only consider
uniform sampling points).}
}
\value{
An estimation of L0.
}
\description{
Perform the estimation of L0.
}
