% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_risk.R
\name{estimate.risk}
\alias{estimate.risk}
\title{Perform the estimation of the risk on a set of curves.}
\usage{
estimate.risk(curves, curves_estim, t0 = 0.5)
}
\arguments{
\item{curves}{List of (true) curves.}

\item{curves_estim}{List of estimated curves.}

\item{t0}{The time point where to compute the risk}
}
\value{
List with the mean and max residual squared error in t0.
}
\description{
Perform the estimation of the risk on a set of curves.
}
