% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_risk.R
\name{estimate.risks}
\alias{estimate.risks}
\title{Perform the estimation of the risk on a set of curves along the points.}
\usage{
estimate.risks(curves, curves_estim)
}
\arguments{
\item{curves}{List of (true) curves.}

\item{curves_estim}{List of estimated curves.}
}
\value{
List with the mean of and max the integrated residual squared error.
}
\description{
Perform the estimation of the risk on a set of curves along the points.
}
