% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.sigma.R
\name{estimate.sigma}
\alias{estimate.sigma}
\title{Perform the estiamtion of the standard deviation of the noise}
\usage{
estimate.sigma(curve)
}
\arguments{
\item{curve}{List of two vectors representing a curve:
- $t Sampling points
- $x Observed points}
}
\value{
The estimated standard deviation
}
\description{
Perform the estiamtion of the standard deviation of the noise
}
