% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_b.R
\name{estimate_b_mean}
\alias{estimate_b_mean}
\title{Perform an estimation of the bandwidth for the estimation of the mean}
\usage{
estimate_b_mean(
  data,
  t0,
  H0 = 0.5,
  L0 = 1,
  sigma = 0,
  variance = 0,
  grid = lseq(0.001, 0.1, length.out = 101),
  nb_obs_minimal = 2
)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{H0}{Numeric, an estimation of \eqn{H_0}.}

\item{L0}{Numeric, an estimation of \eqn{L_0}.}

\item{sigma}{Numeric, an estimation of \eqn{\sigma}.}

\item{variance}{Numeric, an estimation of the variance at \eqn{t_0}.}

\item{grid}{Vector, sequence of bandwidth of which the minimization of the
risk is performed. The default is a logarithmic sequence from 0.001 to 0.1.}

\item{nb_obs_minimal}{Numeric, minimal number of points, in the neighborhood,
to be considered in the computation of the mean. The default is 2.}
}
\value{
Numeric, an estimation of the bandwidth.
}
\description{
This function performs an estimation of the bandwidth for a univariate kernel
regression estimator defined over continuous data using the method of
\cite{add ref}. An estimation of \eqn{H_0}, \eqn{L_0} and \eqn{\sigma} have
to be provided to estimate the bandwidth.
}
\seealso{
Other estimate bandwidth: 
\code{\link{estimate_b_curves_list}()},
\code{\link{estimate_b_curves}()},
\code{\link{estimate_b_cv}()},
\code{\link{estimate_b_mean_list}()}
}
\concept{estimate bandwidth}
