% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_moment.R
\name{estimate_covariance}
\alias{estimate_covariance}
\title{Perform an estimation of the covariance}
\usage{
estimate_covariance(curves, U, b, h, t0_list = NULL)
}
\arguments{
\item{curves}{A list, where each element represents a real curve. Each curve
have to be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{A vector of numerics, sampling points at which estimate the curve.}

\item{b}{A list of bandwidth for the first smoothing}

\item{h}{A list of bandwidth for the second smoothing}

\item{t0_list}{A list of time at which the bandwidths have been estimated.
(default = NULL)}
}
\value{
A matrix of shape (length(U), length(U)) which is an estimation of
the covariance matrix.
}
\description{
This function performs the estimation of the covariance of a set of curves.
}
\examples{
df <- generate_fractional_brownian(N = 10, M = 300, H = 0.5, sigma = 0.05)
estimate_covariance(df, U = seq(0, 1, length.out = 11), b = 0.01, h = 0.05)
}
