% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_H0.R
\name{estimate_H0}
\alias{estimate_H0}
\title{Perform an estimation of \eqn{H_0}}
\usage{
estimate_H0(data)
}
\arguments{
\item{data}{An array, resulting from the presmoothing function.}
}
\value{
Numeric, an estimation of H0.
}
\description{
This function performs an estimation of \eqn{H_0} used for the estimation of
the bandwidth for a univariate kernel regression estimator defined over
continuous domains data using the method of \cite{Golovkine et al. (2021)}.
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\seealso{
Other estimate \eqn{H_0}: 
\code{\link{estimate_H0_deriv_list_old}()},
\code{\link{estimate_H0_deriv_list}()},
\code{\link{estimate_H0_deriv_old}()},
\code{\link{estimate_H0_deriv}()},
\code{\link{estimate_H0_list_old}()},
\code{\link{estimate_H0_list}()},
\code{\link{estimate_H0_old}()}
}
\concept{estimate \eqn{H_0}}
