% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_H0.R
\name{estimate_H0_deriv_list}
\alias{estimate_H0_deriv_list}
\title{Perform an estimation of \eqn{H_0} given a list of \eqn{t_0} when the curves
are derivables}
\usage{
estimate_H0_deriv_list(data, t0_list, gamma = 0.5, Gamma = 1)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{t0_list}{Vector, the sampling point at which we estimate \eqn{H0}. It
corresponds to \eqn{t_2} in the equation (10) of the paper.}

\item{gamma}{Numeric, constant \eqn{\gamma} used in the theorem 1 in the paper.
Should be between 0 and 1.}

\item{Gamma}{Numeric, constant \eqn{\Gamma} used in the theorem 1 in the paper.
Should be striclty greater than 0.}
}
\value{
A vector of numeric, an estimation of \eqn{H_0} at each \eqn{t_0}.
}
\description{
This function performs an estimation of \eqn{H_0} used for the estimation of
the bandwidth for a univariate kernel regression estimator defined over
continuous domains data using the method of \cite{Golovkine et al. (2021)}
in the case the curves are derivables.
}
\examples{
df <- generate_integrate_fractional_brownian(N = 1000, M = 300,
                                             H = 0.5, sigma = 0.01)
H0 <- estimate_H0_deriv_list(df, t0_list = 0.5)
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\seealso{
Other estimate \eqn{H_0}: 
\code{\link{estimate_H0_deriv_list_old}()},
\code{\link{estimate_H0_deriv_old}()},
\code{\link{estimate_H0_deriv}()},
\code{\link{estimate_H0_list_old}()},
\code{\link{estimate_H0_list}()},
\code{\link{estimate_H0_old}()},
\code{\link{estimate_H0}()}
}
\concept{estimate \eqn{H_0}}
