% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_risk.R
\name{estimate_int_risk}
\alias{estimate_int_risk}
\title{Perform an estimation of the risk on one curve along the sampling points.}
\usage{
estimate_int_risk(curves, curves_estim)
}
\arguments{
\item{curves}{A list, where each element represents a real curve. Each curve
have to be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{curves_estim}{A list, where each element represents an estimated curve.
Each curve have to be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The estimated points.
}}
}
\value{
Numeric, the integrated mean squared error
}
\description{
This function performs the estimation of the risk on one curve along
the sampling points. Both the real and estimated curve have to be sampled on
the same grid.
}
\details{
Actually, one risk is computed. They are defined as:
\deqn{IntRSE = \int(X_n(t) - \hat{X}_n(t))^2dt}
}
\examples{
\dontrun{
 X <- generate_fractional_brownian(N = 1000, M = 300, H = 0.5, sigma = 0.05)
 X_smoothed <- smooth_curves(X)$smooth
 estimate_int_risk(X[[1]], X_smoothed[[1]])
}
}
