% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_k0.R
\name{estimate_k0_oracle}
\alias{estimate_k0_oracle}
\title{Perform the estimation of the oracle k0.}
\usage{
estimate_k0_oracle(M, H)
}
\arguments{
\item{M}{Mean number of sampling points per curve.}

\item{H}{Estimation of the regularity of the curves.}
}
\value{
The oracle k0.
}
\description{
Perform the estimation of the oracle k0.
}
