% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_L0.R
\name{estimate_L0_list}
\alias{estimate_L0_list}
\title{Perform the estimation of \eqn{L_0} given a list of \eqn{t_0}}
\usage{
estimate_L0_list(
  data,
  t0_list,
  H0_list,
  k0_list = 2,
  sigma = NULL,
  density = FALSE
)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{t0_list}{A vetor of numerics, the sampling points at which we estimate
\eqn{H0}. We will consider the \eqn{8k0 - 7} nearest points of \eqn{t_0} for
the estimation of \eqn{H_0} when \eqn{\sigma} is unknown.}

\item{H0_list}{A vector of numerics, an estimation of \eqn{H_0} at every
\eqn{t_0} given in \code{t0_list}.}

\item{k0_list}{A vector of numerics, the number of neighbors of \eqn{t_0} to
consider. Should be set as \eqn{k0 = M * exp(-log(log(M))^2)}.}

\item{sigma}{Numeric, true value of sigma. Can be NULL.}

\item{density}{Logical, do the sampling points have a uniform distribution?
(default is FALSE)}
}
\value{
A vector of numerics, an estimation of \eqn{L_0} at each \eqn{t_0}.
}
\description{
This function performs an estimation of \eqn{H_0} used for the estimation of
the bandwidth for a univariate kernel regression estimator defined over
continuous domains data using the method of \cite{add ref}.
}
\examples{
df <- generate_fractional_brownian(N = 1000, M = 300, H = 0.5, sigma = 0.05)
L0 <- estimate_L0_list(df, t0_list = 0.5, H0_list = 0.5)

df_piece <- generate_piecewise_fractional_brownian(N = 1000, M = 300, 
                                                   H = c(0.2, 0.5, 0.8), 
                                                   sigma = 0.05)
L0 <- estimate_L0_list(df_piece, t0_list = c(0.15, 0.5, 0.85), 
                       H0_list = c(0.2, 0.5, 0.8), k0 = 6)
}
\seealso{
Other estimate \eqn{L_0}: 
\code{\link{estimate_L0}()}
}
\concept{estimate \eqn{L_0}}
