% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_moment.R
\name{estimate_LOO_mean}
\alias{estimate_LOO_mean}
\title{Perform a leave-one-out mean curve}
\usage{
estimate_LOO_mean(curves, b, n, t0_list = NULL)
}
\arguments{
\item{curves}{A list, where each element represents a real curve. Each curve
have to be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{b}{A list of bandwidth}

\item{n}{The curves to remove from the estimation}

\item{t0_list}{A list of time at which the bandwidths have been estimated.
(default = NULL)}
}
\value{
A vector representing the LOO mean curve.
}
\description{
This function performs the computation of a leave-one-out mean curve.
}
