% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_moment.R
\name{estimate_mean}
\alias{estimate_mean}
\title{Perform an estimation of the leave-one-out mean curve}
\usage{
estimate_mean(curves, U, b, t0_list = NULL)
}
\arguments{
\item{curves}{A list, where each element represents a real curve. Each curve
have to be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{A vector of numerics, sampling points at which estimate the curve.}

\item{b}{A list of bandwidth}

\item{t0_list}{A list of time at which the bandwidths have been estimated.
(default = NULL)}
}
\value{
A vector representing the mean curve.
}
\description{
This function performs the estimation of the leave-one-out mean curve of a
set of curves.
}
