% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sigma.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Perform an estimation of the standard deviation of the noise}
\usage{
estimate_sigma(data, t0 = 0.5, k0 = 2)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{t0}{Numeric, the sampling point at which we estimate \eqn{\sigma}. We will
consider the \eqn{8k0 - 7} nearest points of \eqn{t_0} for the estimation of
\eqn{\sigma}}

\item{k0}{Numeric, the number of neighbors of \eqn{t_0} to consider. Should
be set as \eqn{k0 = M * exp(-log(log(M))^2)}.}
}
\value{
Numeric, an estimation of the standard deviation of the noise.
}
\description{
This function performs an estimation of the standard deviation of the noise
in the curves. Based on \cite{add ref}, the following formula is used:
\deqn{\hat{\sigma^2} = \frac{1}{N}\sum_{n = 1}^{N} 
      \frac{1}{2(M_n - 1)}\sum_{l = 2}^{M_n}(Y_{n, (l)} - Y_{n, (l-1)})^2}
}
\examples{
df <- generate_fractional_brownian(N = 1000, M = 300, H = 0.5, sigma = 0.05)
estimate_sigma(df, t0 = 0.5, k0 = 14)
}
