% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sigma.R
\name{estimate_sigma_list}
\alias{estimate_sigma_list}
\title{Perform an estimation of the noise given a list of \eqn{t_0}.}
\usage{
estimate_sigma_list(data, t0_list, k0_list)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{t0_list}{A vector of numerics, the sampling points at which we estimate
\eqn{H0}. We will consider the \eqn{8k0 - 7} nearest points of \eqn{t_0} for
the estimation of \eqn{H_0} when \eqn{\sigma} is unknown.}

\item{k0_list}{A vector of numerics, the number of neighbors of \eqn{t_0} to
consider. Should be set as \deqn{k0 = M * exp(-(log(log(M))**2))}. We can set a
different \eqn{k_0}, but in order to use the same for each \eqn{t_0}, just
put a unique numeric.}
}
\value{
A list, an estimation of sigma at different \eqn{t_0}.
}
\description{
This function performs an estimation of the standard deviation of the noise
at different times.
}
\examples{
df <- generate_fractional_brownian(N = 1000, M = 300, H = 0.5, sigma = 0.05)
estimate_sigma_list(df, t0_list = c(0.25, 0.5, 0.75), k0 = 14)
}
