% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_H0.R
\name{estimate_var}
\alias{estimate_var}
\title{Perform an estimation of \eqn{var(X_{t_0)}}}
\usage{
estimate_var(data)
}
\arguments{
\item{data}{A list of array, resulting from the presmoothing function.}
}
\value{
List, estimation of the variance at each \eqn{t_0}.
}
\description{
This function performs an estimation of \eqn{var(X_{t_0})} used for the
estimation of the bandwidth for the mean and the covariance by a univariate
kernel regression estimator.
}
