% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{fractional_brownian_trajectory}
\alias{fractional_brownian_trajectory}
\title{Generate fractional Brownian motion with a random noise}
\usage{
fractional_brownian_trajectory(M, H, sigma, pdf = NULL, L = 1)
}
\arguments{
\item{M}{An integer, expected number of points in the trajectory.
The number of points follows a Poisson distribution with mean \eqn{M}.}

\item{H}{Numeric, Hurst coefficient. \eqn{0 < H < 1}}

\item{sigma}{A vector of numerics, standard deviation of the noise to add to
the fractional Brownian motion.}

\item{pdf}{Function, probability density function for the sampling points.}

\item{L}{Numeric, multiplicative constant.}
}
\value{
A tibble containing the following elements:
\itemize{
\item \strong{...1} The sampling points
\item \strong{...2} The true trajectory
\item \strong{...3} The trajectory contaminated by noise with standard
deviation \eqn{\sigma}
}
}
\description{
This function generates a realization of a fractional Brownian motion with
random noise. The increments of a fractional Brownian motion are not
independent. A fractional Brownian motion is characterized by a parameter
\eqn{H}, named Hurst coefficient. We define it on \eqn{[0, 1]}.
}
