% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_fractional_brownian}
\alias{generate_fractional_brownian}
\title{Generate a list of fractional Brownian trajectory.}
\usage{
generate_fractional_brownian(
  N = 100,
  M = 10,
  H = 0.5,
  sigma = 0.05,
  pdf = NULL,
  L = 1
)
}
\arguments{
\item{N}{An integer, number of curves to simulate.}

\item{M}{An integer, expected number of points in the trajectory.
The number of points follows a Poisson distribution with mean \eqn{M}.}

\item{H}{Numeric, Hurst coefficient. \eqn{0 < H < 1}}

\item{sigma}{A vector of numerics, standard deviation of the noise to add to
the fractional Brownian motion.}

\item{pdf}{Function, probability density function for the sampling points.}

\item{L}{Numeric, multiplicative constant.}
}
\value{
A tibble containing the following elements:
\itemize{
\item \strong{...1} The sampling points
\item \strong{...2} The true trajectory
\item \strong{...3} The trajectory contaminated by noise with standard
deviation \eqn{\sigma}
}
}
\description{
This function generates a list of realizations of a fractional Brownian
motion with random noise. The increments of a fractional Brownian motion
are not independent. A fractional Brownian motion is characterized by a
parameter \eqn{H}, named Hurst coefficient. We define it on \eqn{[0, 1]}.
}
\examples{
generate_fractional_brownian(100, 50, 0.7, 0.1)
generate_fractional_brownian(100, 50, 0.5, 0.05, pdf = rnorm)
}
