% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kernel}
\alias{kernel}
\title{Compute different kernels}
\usage{
kernel(u, type = 1)
}
\arguments{
\item{u}{Vector of numeric, points to estimate the kernel.}

\item{type}{Integer, used kernel. If 1, uniform kernel, If 2, Epanechnikov
kernel. If 3, biweight kernel.}
}
\value{
Vector of numeric
}
\description{
Compute different kernels
}
