% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2cai}
\alias{list2cai}
\title{Convert comprehensive lists into \code{ssfcov} objects.}
\usage{
list2cai(data)
}
\arguments{
\item{data}{list, where each element represents a curve. Each curve is defined
as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}
}
\value{
A dataframe where the columns are:
\itemize{
\item \strong{$time} The observation times.
\item \strong{x} The observation values.
\item \strong{subject} The observation id.
}
}
\description{
Convert comprehensive lists into \code{ssfcov} objects.
}
\references{
ssfcov, R package. See Cai and Yuan, Nonparametric Covariance
Function Estimation for Functional and Longitudinal Data - 2010.
}
