% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2funData}
\alias{list2funData}
\title{Convert comprehensive lists into \code{funData::funData} objects.}
\usage{
list2funData(data_list)
}
\arguments{
\item{data_list}{A list, where each element represents a curve. Each curve is
defined as list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}
}
\value{
An object of the class \code{funData::funData}
}
\description{
We assume that we \strong{know} that the curves are on the same interval.
}
