% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2irregFunData}
\alias{list2irregFunData}
\title{Convert comprehensive lists into \code{funData::irregFunData} objects.}
\usage{
list2irregFunData(data_list)
}
\arguments{
\item{data_list}{A list, where each element represents a curve. Each curve is
defined as list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}
}
\value{
An object of the class \code{funData::irregFunData}
}
\description{
Convert comprehensive lists into \code{funData::irregFunData} objects.
}
