% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lseq}
\alias{lseq}
\title{Generate logarithmic spaced sequence}
\usage{
lseq(from = 1, to = 100, length.out = 51)
}
\arguments{
\item{from}{Numeric}

\item{to}{Numeric}

\item{length.out}{Numeric}
}
\value{
Vector
}
\description{
Generate logarithmic spaced sequence
}
