% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multiFunData2list}
\alias{multiFunData2list}
\title{Convert \code{multiFunData} objects into right format for this package}
\usage{
multiFunData2list(data, norm = TRUE)
}
\arguments{
\item{data}{An object of the class \code{funData::multiFunData}}

\item{norm}{Boolean, if TRUE, the sampling are normalized on \eqn{[0, 1]}}
}
\value{
A list, where each element represents a curve. Each curve is defined
as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}
@export
}
\description{
Convert \code{multiFunData} objects into right format for this package
}
