% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{neighbors}
\alias{neighbors}
\title{Test whether the observation points are in the neighborhood}
\usage{
neighbors(t, t0, h, k0)
}
\arguments{
\item{t}{Vector of numeric}

\item{t0}{Numeric}

\item{h}{Numeric}

\item{k0}{Numeric}
}
\value{
Vector of boolean
}
\description{
Test whether the observation points are in the neighborhood
}
