% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{piecewise_fractional_brownian_trajectory}
\alias{piecewise_fractional_brownian_trajectory}
\title{Generate piecewise fractional Brownian motion with a random noise}
\usage{
piecewise_fractional_brownian_trajectory(M, H, sigma, pdf = NULL)
}
\arguments{
\item{M}{An integer, expected number of points in the trajectory.
The number of points follows a Poisson distribution with mean \eqn{M}.}

\item{H}{A vector of numeric, Hurst coefficients. \eqn{0 < H_k < 1}}

\item{sigma}{A vector of numeric, standard deviation of the noise to add to
the piecewise fractional Brownian motion. Should have the length of H. It
adds heteroscedastic noise to the data.}

\item{pdf}{A function for the generation of the sampling points.}
}
\value{
A tibble containing the following elements:
\itemize{
\item \strong{...1}: The sampling points
\item \strong{...2} The true trajectory
\item \strong{...3} The trajectory contaminated by noise with standard
deviation \eqn{\sigma}
}
}
\description{
This function generates a realization of a piecewise fractional Brownian motion
with random noise. A piecewise fractional Brownian motion is defined by a non
constant Hurst parameter along the sampling points. We observe the process at
regularly spaced time \eqn{t_i = \frac{i}{M_n}}, where \eqn{i = 0, \dots, M_n}.
We define a segmentation \eqn{\tau = (\tau_k)_{k=0, \dots, K+1}}, with
\eqn{0 = \tau_0 < \tau_1 < \dots < \tau_{K} < \tau_{K+1} = 1}. So, on the
interval \eqn{[\tau_k, \tau_{k+1}]}, for \eqn{k = 0, \dots, K}, the process
is a fractional Brownian motion with Hurst parameter \eqn{H_k}.
}
