% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_H0.R
\name{presmoothing}
\alias{presmoothing}
\title{Perform a presmoothing of the data}
\usage{
presmoothing(data, t0_list = 0.5, gamma = 0.5, order = 1, drv = 0, degree = 0)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{t0_list}{List, the sampling point around which we presmooth the data.}

\item{gamma}{Numeric, constant \eqn{\gamma} used in the theorem 1 in the paper.
Should be between 0 and 1.}

\item{order}{Integer, default=0. Regularity of the input data.}

\item{drv}{Integer, default=0. Order of derivative to be estimated.}
}
\value{
List of array which contains the smoothed data.
}
\description{
This function performs a presmoothing of the data by local linear smoother.
}
