% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_curves.R
\name{smooth.curves}
\alias{smooth.curves}
\title{Perform a non-parametric smoothing of a set of curves.}
\usage{
smooth.curves(data, U = NULL, t0 = 0.5, k0 = 2, K = "epanechnikov")
}
\arguments{
\item{data}{A list of curves to smooth. Each entry of the list should have
two elements:
\itemize{
\item $t which correspond to the time we observed the curve.
\item $x which correspond to the values observed.
}}

\item{U}{Vector of points for the estimation.
If NULL, smooth the curves on the same grid than they are observed.}

\item{t0}{The starting time for the estimation of H0. We consider the 8k0 - 7
nearest points of t0 for the estimation of H0 when sigma is unknown. Can be
a list.}

\item{k0}{For the computation of the gap between the different observations.
Should be set as k0 = M / max(8, log(M)).}

\item{K}{Kernel used for the estimation.
\itemize{
\item epanechnikov (default)
}}
}
\value{
A list of same size of \code{data} containing the smoothed curves.
}
\description{
Perform a non-parametric smoothing of a set of curves.
}
