% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.curves.R
\name{smooth.curves.regularity}
\alias{smooth.curves.regularity}
\title{Perform a non-parametric smoothing of a set of curves when the regularity is
larger than 1.}
\usage{
smooth.curves.regularity(
  data,
  U = NULL,
  t0 = 0.5,
  k0 = 2,
  K = "epanechnikov",
  eps = 0.1
)
}
\arguments{
\item{data}{A list of curves to smooth. Each entry of the list should have
two elements:
\itemize{
\item $t which corresponds to the time we observed the curve.
\item $x which corresponds to the values observed.
We assume that the underlying regularity of the curve is larger than 1.
}}

\item{U}{Vector of points for the estimation.
If NULL, smooth the curves on the same grid than they are observed.}

\item{t0}{The starting time for the estimation of H0. We consider the 8k0 - 7
nearest points of t0 for the estimation of H0 when sigma is unknown.}

\item{k0}{For the computation of the gap between the different observations.
Shouldd be set as k0 = M / max(8, log(M)).}

\item{K}{Kernel used for the estimation.
\itemize{
\item epanechnikov (default)
}}

\item{eps}{Precision parameter.}
}
\value{
A list with two entries:
\itemize{
\item parameter which contains the estimation of sigma, H0, L0 and b.
\item smooth which is a list of the same size than \code{data} containing the
smoothed curves.
}
}
\description{
Perform a non-parametric smoothing of a set of curves when the regularity is
larger than 1.
}
