% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_curves.R
\name{smooth_curves_regularity}
\alias{smooth_curves_regularity}
\title{Perform a non-parametric smoothing of a set of curves when the regularity is
larger than 1}
\usage{
smooth_curves_regularity(
  data,
  U = NULL,
  t0 = 0.5,
  k0 = 2,
  K = "epanechnikov",
  eps = 0.1
)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{U}{A vector of numerics, sampling points at which estimate the curves.
If NULL, the sampling points for the estimation are the same than the
observed ones.}

\item{t0}{Numeric, the sampling point at which we estimate \eqn{H0}. We will
consider the \eqn{8k0 - 7} nearest points of \eqn{t_0} for the estimation of
\eqn{H_0} when \eqn{\sigma} is unknown.}

\item{k0}{Numeric, the number of neighbors of \eqn{t_0} to consider. Should
be set as \eqn{k0 = M * exp(-log(log(M))^2)}.}

\item{K}{Character string, the kernel used for the estimation:
\itemize{
\item epanechnikov (default)
\item uniform
\item beta
}}

\item{eps}{Numeric, precision parameter. It is used to control how much larger
than 1, we have to be in order to consider to have a regularity larger than 1
(default to 0.1).}
}
\value{
A list, which contains two elements. The first one is a list which
contains the estimated parameters:
\itemize{
\item \strong{sigma} An estimation of the standard deviation of the noise
\item \strong{H0} An estimation of \eqn{H_0}
\item \strong{L0} An estimation of \eqn{L_0}
\item \strong{b} An estimation of the bandwidth
}
The second one is another list which contains the estimation of the curves:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The estimated points.
}
}
\description{
This function performs a non-parametric smoothing of a set of curves using the
Nadaraya-Watson estimator when the regularity of the underlying curves is
larger than 1. The bandwidth is estimated using the method from
\cite{add ref}. In the case of a regularly larger than 1, we currently
assume that the regularly is the same all over the curve.
}
\examples{
df <- generate_integrate_fractional_brownian(N = 1000, M = 300, 
                                              H = 0.5, sigma = 0.01)
df_smooth <- smooth_curves_regularity(df, U = seq(0, 1, length.out = 101), 
                                      t0 = 0.5, k0 = 14)
}
