% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rkhs.R
\name{ssf_covariance}
\alias{ssf_covariance}
\title{Estimation of the covariance with spline smoothing}
\usage{
ssf_covariance(
  data,
  t = seq(0, 1, length.out = 101),
  s = NULL,
  nbasis = 5,
  centered = FALSE,
  noDiag = TRUE
)
}
\arguments{
\item{data}{A list, where each element represents a real curve. Each curve
have to be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{t}{Vector, contains the times where the covariance is computed.}

\item{s}{Vector, contains the times where the covariance is computed on the
other axis. IF NULL, set to be equal to t.}

\item{nbasis}{Integer, number of basis to use for the splines.}

\item{centered}{Boolean, Do the data are centered? (default = FALSE)}

\item{noDiag}{Boolean, Remove the diagonal of the covariance in the data?
(dafault = TRUE)}
}
\description{
This function performs the estimation of the covariance function based on the
paper by Cai and Yuan.
}
\references{
Cai and Yuan, Nonparametric Covariance Function Estimation for
Functional and Longitudinal Data, 2010.
}
