% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rkhs.R
\name{ssf_mean}
\alias{ssf_mean}
\title{Estimation of the mean with spline smoothing}
\usage{
ssf_mean(data, time)
}
\arguments{
\item{data}{A list, where each element represents a real curve. Each curve
have to be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points
}}

\item{time}{A vector containing the times where the mean is computed.}
}
\value{
A vector representing the mean curve.
}
\description{
This function performs the estimation of the mean function based on the paper
by Cai and Yuan who show the optimal estimation of the mean function under
both common and independent designs.
}
\references{
Cai and Yuan, Optimal Estimation of the mean function based on
discretely sampled functional data: phase transition, 2011,
The Annals of Statistics
}
