% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canadian_temperature_daily}
\alias{canadian_temperature_daily}
\title{Daily canadian temperature}
\format{An object of class \code{list} of length 35.}
\usage{
canadian_temperature_daily
}
\description{
Daily canadian temperature
}
\keyword{datasets}
