% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.curve.R
\name{estimate.curve}
\alias{estimate.curve}
\title{Perform the smoothing of individuals curves}
\usage{
estimate.curve(curve, U, b)
}
\arguments{
\item{curve}{List of two vectors representing a curve:
- $t Sampling points
- $x Observed points}

\item{U}{Vector of points for the estimation}

\item{b}{Bandwith}
}
\value{
List of two vectors representaing the estimated curve:
- $t Sampling points
- $x Predicted points
}
\description{
Perform the smoothing of individuals curves
}
