% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_b.R
\name{estimate_b_cv}
\alias{estimate_b_cv}
\title{Perform an estimation of the bandwidth using least-squares cross validation}
\usage{
estimate_b_cv(data)
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}
}
\value{
Numeric, an estimation of the bandwidth.
}
\description{
The function performs an estimation of the bandwidth for a univariate kernel
regression estimator defined over continuous data using least-square cross
validation for each curve and return the average bandwidth among them.
}
\examples{
df <- denoisr::generate_fractional_brownian(N = 5, M = 300, 
                                            H = 0.5, sigma = 0.05)
b <- estimate_b_cv(df)
}
\seealso{
\code{\link[np]{npregbw}}

Other estimate bandwidth: 
\code{\link{estimate_b_curves_list}()},
\code{\link{estimate_b_curves}()},
\code{\link{estimate_b_mean_list}()},
\code{\link{estimate_b_mean}()}
}
\concept{estimate bandwidth}
