% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_b.R
\name{estimate_b_list}
\alias{estimate_b_list}
\title{Perform an estimation of the bandwidth given a list of \eqn{H_0} and \eqn{L_0}}
\usage{
estimate_b_list(data, H0_list, L0_list, sigma = 0, K = "epanechnikov")
}
\arguments{
\item{data}{A list, where each element represents a curve. Each curve have to
be defined as a list with two entries:
\itemize{
\item \strong{$t} The sampling points
\item \strong{$x} The observed points.
}}

\item{H0_list}{A vector of numerics, estimations of \eqn{H_0}.}

\item{L0_list}{A vector of numerics, estimations of \eqn{L_0}.}

\item{sigma}{Numeric, an estimation of \eqn{\sigma}.}

\item{K}{Character string, the kernel used for the estimation:
\itemize{
\item epanechnikov (default)
\item beta
\item uniform
}}
}
\value{
A vector of numerices, estimations of the bandwidth.
}
\description{
This function performs an estimation of the bandwidth for a univariate kernel
regression estimator defined over continuous data using the method of
\cite{add ref}. An estimation of \eqn{H_0}, \eqn{L_0} and \eqn{\sigma} have
to be provided to estimate the bandwidth.
}
\examples{
df <- denoisr::generate_fractional_brownian(N = 1000, M = 300, 
                                            H = 0.5, sigma = 0.05)
b <- estimate_b_list(df, H0_list = 0.5, L0_list = 1, sigma = 0.05)

df_piece <- generate_piecewise_fractional_brownian(N = 1000, M = 300, 
                                                   H = c(0.2, 0.5, 0.8), 
                                                   sigma = 0.05)
b <- estimate_b_list(df_piece, H0_list = c(0.2, 0.5, 0.8), L0_list = c(1, 1, 1),
                     sigma = 0.1, K = 'epanechnikov')
}
\seealso{
Other estimate bandwidth: 
\code{\link{estimate_b_cv}()},
\code{\link{estimate_b}()}
}
\concept{estimate bandwidth}
