% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_H0.R
\name{estimate_H0_list}
\alias{estimate_H0_list}
\title{Perform an estimation of \eqn{H_0} given a list of \eqn{t_0}}
\usage{
estimate_H0_list(data)
}
\arguments{
\item{data}{A list of array, resulting from the presmoothing function.}
}
\value{
A vector of numeric, an estimation of \eqn{H_0} at each \eqn{t_0}.
}
\description{
This function performs an estimation of \eqn{H_0} used for the estimation of
the bandwidth for a univariate kernel regression estimator defined over
continuous domains data using the method of \cite{Golovkine et al. (2020)}.
}
\examples{
df <- generate_fractional_brownian(N = 1000, M = 300, H = 0.5, sigma = 0.05)
df_smooth <- presmoothing(df, t0_list = 0.5, gamma = 0.5, order = 1,
                          drv = 0, degree = 0)
H0 <- estimate_H0_list(df_smooth)

df_piece <- generate_piecewise_fractional_brownian(N = 1000, M = 300, 
                                                   H = c(0.2, 0.5, 0.8), 
                                                   sigma = 0.05)
df_smooth <- presmoothing(df, t0_list = c(0.2, 0.5, 0.8), gamma = 0.5,
                          order = 1, drv = 0, degree = 0)
H0 <- estimate_H0_list(df_smooth)
}
\references{
Golovkine S., Klutchnikoff N., Patilea V. (2021) - Adaptive
estimation of irregular mean and covariance functions.
}
\seealso{
Other estimate \eqn{H_0}: 
\code{\link{estimate_H0_deriv_list_old}()},
\code{\link{estimate_H0_deriv_list}()},
\code{\link{estimate_H0_deriv_old}()},
\code{\link{estimate_H0_deriv}()},
\code{\link{estimate_H0_list_old}()},
\code{\link{estimate_H0_old}()},
\code{\link{estimate_H0}()}
}
\concept{estimate \eqn{H_0}}
