% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_k0.R
\name{estimate_k0_pilot}
\alias{estimate_k0_pilot}
\title{Perform the estimation of the pilot k0.}
\usage{
estimate_k0_pilot(M)
}
\arguments{
\item{M}{Mean number of samping points per curve.}
}
\value{
The pilot k0.
}
\description{
Perform the estimation of the pilot k0.
}
