% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sigma.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{Perform the estimation of sigma}
\usage{
estimate_sigma(data)
}
\arguments{
\item{data}{List of curves to estimate by kernel regression.}
}
\value{
An estimation of sigma.
}
\description{
Perform the estimation of sigma
}
