% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fractional_brownian}
\alias{fractional_brownian}
\title{Fractional brownian trajectories}
\format{An object of class \code{list} of length 1000.}
\usage{
fractional_brownian
}
\description{
Fractional brownian trajectories
}
\keyword{datasets}
