% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{piecewise_fractional_brownian}
\alias{piecewise_fractional_brownian}
\title{Piecewise fractional brownian trajectories}
\format{An object of class \code{list} of length 1000.}
\usage{
piecewise_fractional_brownian
}
\description{
Piecewise fractional brownian trajectories
}
\keyword{datasets}
